add client bundle and static assets
