+++ public/assets/App.js
@@ -1,0 +1,5 @@
+import { jsx as _jsx, jsxs as _jsxs } from "https://esm.sh/react@18.3.1/jsx-runtime";
+import React from "https://esm.sh/react@18.3.1";
+export function App() {
+    return (_jsxs("html", { lang: "en", children: [_jsxs("head", { children: [_jsx("meta", { charSet: "utf-8" }), _jsx("meta", { name: "viewport", content: "width=device-width, initial-scale=1" }), _jsx("title", { children: "Falcon Systems \u2013 Teaming with AI" }), _jsx("link", { rel: "stylesheet", href: "/assets/index.css" })] }), _jsxs("body", { className: "min-h-screen bg-gradient-to-b from-white to-slate-50 text-slate-900 transition-colors dark:from-slate-900 dark:to-black dark:text-slate-100", children: [_jsxs("header", { className: "py-6 px-4 max-w-6xl mx-auto flex items-center", children: [_jsx("h1", { className: "text-3xl font-bold flex-1", children: "Falcon Systems" }), _jsx("button", { id: "theme-toggle", className: "px-3 py-1 rounded-full border border-slate-900 text-slate-900 dark:border-slate-100 dark:text-slate-100", type: "button", children: "Toggle theme" })] }), _jsxs("main", { className: "px-4 max-w-6xl mx-auto space-y-16", children: [_jsxs("section", { id: "home", children: [_jsxs("h2", { className: "text-4xl font-semibold", children: ["A practical framework for ", "", _jsx("span", { id: "rotating-tagline", className: "bg-gradient-to-r from-indigo-500 via-purple-500 to-pink-500 text-transparent bg-clip-text transition-opacity duration-500 inline-block", children: "teaming with AI" })] }), _jsx("p", { className: "text-slate-600 mt-3", children: "Falcon Systems helps organizations adopt AI safely and effectively\u2014through an opinionated framework, hands\u2011on curriculum, and a managed, sandboxed environment to learn by doing." }), _jsxs("div", { className: "mt-4 flex gap-3", children: [_jsx("a", { href: "#framework", className: "px-4 py-2 rounded-full bg-slate-900 text-white", children: "Explore the framework" }), _jsx("a", { href: "#contact", className: "px-4 py-2 rounded-full border border-slate-900 text-slate-900", children: "Get in touch" })] })] }), _jsxs("section", { id: "framework", children: [_jsx("h2", { className: "text-3xl font-semibold", children: "The Falcon Framework" }), _jsxs("div", { className: "grid md:grid-cols-4 gap-4 mt-4", children: [_jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Roles" }), _jsx("p", { className: "text-slate-600 text-sm", children: "Define responsibilities across people & AI." })] }), _jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Interfaces" }), _jsx("p", { className: "text-slate-600 text-sm", children: "Design stable, testable interactions." })] }), _jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Guardrails" }), _jsx("p", { className: "text-slate-600 text-sm", children: "Mitigate leakage and risk." })] }), _jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Evaluation" }), _jsx("p", { className: "text-slate-600 text-sm", children: "Measure quality, risk, and ROI." })] })] })] }), _jsxs("section", { id: "offerings", children: [_jsx("h2", { className: "text-3xl font-semibold", children: "Service offerings" }), _jsxs("div", { className: "grid md:grid-cols-3 gap-4 mt-4", children: [_jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Foundations" }), _jsx("p", { className: "text-slate-600 text-sm", children: "Core principles, prompt patterns, live labs." })] }), _jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Enablement" }), _jsx("p", { className: "text-slate-600 text-sm", children: "From pilots to practice; rollout playbooks." })] }), _jsxs("div", { className: "p-4 border rounded-2xl", children: [_jsx("strong", { children: "Managed Sandbox" }), _jsx("p", { className: "text-slate-600 text-sm", children: "Isolated experimentation with guardrails." })] })] })] }), _jsxs("section", { id: "contact", children: [_jsx("h2", { className: "text-3xl font-semibold", children: "Let\u2019s talk" }), _jsxs("form", { method: "post", action: "/api/contact", className: "grid gap-3 max-w-lg mt-4", children: [_jsx("input", { required: true, name: "name", placeholder: "Name", className: "border rounded-xl p-3" }), _jsx("input", { required: true, type: "email", name: "email", placeholder: "Email", className: "border rounded-xl p-3" }), _jsx("input", { name: "company", placeholder: "Company", className: "border rounded-xl p-3" }), _jsx("textarea", { required: true, name: "message", placeholder: "How can we help?", rows: 4, className: "border rounded-xl p-3" }), _jsx("button", { className: "px-4 py-2 rounded-full bg-slate-900 text-white", type: "submit", children: "Submit" })] }), _jsx("p", { className: "text-slate-600 mt-2", children: "Or email: hello@falconsystems.ai" })] })] }), _jsx("footer", { className: "py-10 border-t mt-16 px-4 max-w-6xl mx-auto", children: _jsxs("p", { className: "text-slate-500 text-sm", children: ["\u00A9 ", new Date().getFullYear(), " Falcon Systems. All rights reserved."] }) }), _jsx("script", { type: "module", src: "/assets/client.js" })] })] }));
+}
