+++ public/assets/index.css
@@ -1,0 +1,1 @@
+@import url("https://cdn.jsdelivr.net/npm/tailwindcss@3.4.1/dist/tailwind.min.css");
