+++ wrangler.toml
@@ -1,3 +1,6 @@
 name = "falconsystems-ssr"
 main = "src/worker.tsx"
 compatibility_date = "2024-09-02"
+
+[assets]
+directory = "./public"
